fix: Fix initial API URL by location #235
