+++ src/views/Main.tsx
@@ -1,293 +1,19 @@
-import { ThemeProvider } from "@emotion/react";
-import styled from "@emotion/styled";
 import { useDispatch, useSelector } from "react-redux";
 import { themes } from "../theme/themes";
-import Panel from "../components/Panel/Panel";
-import { Notification } from "../qryn-ui/notifications";
-import SettingsDialog from "../plugins/settingsdialog/SettingsDialog";
 import { UpdateStateFromQueryParams } from "../helpers/UpdateStateFromQueryParams";
-import StatusBar from "../components/StatusBar";
-import QueryHistory from "../plugins/queryhistory";
 import { useMediaQuery } from "react-responsive";
-import MainTabs from "./MainTabs";
 import { setTheme } from "../actions";
-import { useMemo, useState, useEffect, useRef } from "react";
-import { useCookies } from "react-cookie";
-import { useLocation } from "react-router-dom";
-import setDataSources from "./DataSources/store/setDataSources";
-import { setShowDataSourceSetting } from "./Main/setShowDataSourceSetting";
-
-export const MainContainer = styled.div`
-    position: absolute;
-    display: flex;
-    flex-direction: column;
-    height:inherit;
-    width: 100%;
-    flex: 1;
-
-    background-color: ${({theme}: any) => theme.mainBgColor} !important;
-    &::-webkit-scrollbar-corner {
-        background: transparent;
-    }
-    &::-webkit-scrollbar-thumb {
-        border-radius: 5px;
-        background: ${({theme}: any) => theme.scrollbarThumb} !important;
-    }
-    .panels-container {
-        display: flex;
-        background: ${({ theme }: any) => theme.shBgColor};
-        height: calc(100vh - 45px);
-    }
-`;
-
-/**
- *
- * @param {theme, isEmbed, settingsDialogOpen}
- * @returns Mobile View
- */
-
-export function MobileView({ theme, isEmbed, settingsDialogOpen }: any) {
-    return (
-        <ThemeProvider theme={theme}>
-            {!isEmbed && <StatusBar />}
-
-            <MainContainer>
-                <MainTabs />
-            </MainContainer>
-
-            <Notification />
-            <SettingsDialog open={settingsDialogOpen} />
-            <QueryHistory />
-        </ThemeProvider>
-    );
-}
-
-/**
- *
- * @param {theme, isEmbed, isSplit, settingsDialogOpen}
- * @returns Desktop View
- */
-
-export function DesktopView({ theme, isEmbed, isSplit, settingsDialogOpen }: any) {
-    const [height, setHeight] = useState(0);
-    const [widthTotal, setWidthTotal] = useState(0);
-    const [widthLeft, setWidthLeft] = useState(0);
-    const [widthRight, setWidthRight] = useState(0);
-    const [widthLeftPercent, setWidthLeftPercent] = useState(0);
-    const [widthRightPercent, setWidthRightercent] = useState(0);
-    const [minWidth, setMinWidth] = useState(0);
-    const [maxWidth, setMaxWidth] = useState(0);
-    const refTotal: any = useRef(null);
-    useEffect(() => {
-        const widthTotal = refTotal.current.clientWidth;
-        setHeight(refTotal.current.clientHeight);
-        setWidthTotal(refTotal.current.clientWidth);
-        setWidthLeft(widthTotal / (isSplit ? 2 : 1));
-        if (isSplit) {
-            setWidthRight(widthTotal / 2);
-        }
-        const realMinWidth = !isSplit
-            ? widthTotal
-            : widthTotal / 4 > 370
-            ? widthTotal / 4
-            : 370;
-        setMinWidth(realMinWidth);
-        const realMaxWidth = !isSplit ? widthTotal : widthTotal - realMinWidth;
-        setMaxWidth(realMaxWidth);
-    }, [
-        setWidthLeft,
-        setWidthRight,
-        setWidthTotal,
-        setHeight,
-        setMinWidth,
-        setMaxWidth,
-        minWidth,
-        isSplit,
-    ]);
-    useEffect(() => {
-        const widthTotal = refTotal.current.clientWidth;
-        setWidthLeftPercent(widthLeft / widthTotal);
-        if (isSplit) {
-            setWidthRightercent(widthRight / widthTotal);
-        }
-    }, [widthLeft, widthRight]);
-    useEffect(() => {
-        const onWindowResize = () => {
-            const widthTotal = refTotal.current.clientWidth;
-            setWidthTotal(widthTotal);
-            setWidthLeft(widthTotal * widthLeftPercent);
-            if (isSplit) {
-                setWidthRight(widthTotal * widthRightPercent);
-            }
-        };
-        window.addEventListener("resize", onWindowResize);
-        return () => {
-            window.removeEventListener("resize", onWindowResize);
-        };
-    }, [
-        widthTotal,
-        widthLeft,
-        widthRight,
-        widthLeftPercent,
-        widthRightPercent,
-        isSplit,
-    ]);
-
-    return (
-        <ThemeProvider theme={theme}>
-            <MainContainer>
-                {!isEmbed && <StatusBar />}
-                <div className="panels-container" ref={refTotal}>
- 
-                    <Panel name="left" />
-                  
-                    {isSplit && (
-  
-                        <Panel name="right" />
-                     
-                    )}
-                </div>
-            </MainContainer>
-            <SettingsDialog open={settingsDialogOpen} />
-            <QueryHistory />
-        </ThemeProvider>
-    );
-}
-
-// useCookiesAvailable:
-
-export function useCookiesAvailable(urlParams: any) {
-    let cookieAuth = "";
-    let cookiesAvailable = false;
-
-    const hasCookie = useMemo(() => {
-        return urlParams.has("cookie") || false;
-    }, [urlParams]);
-
-    const cookieParam = useMemo(() => {
-        if (hasCookie) {
-            return urlParams.get("cookie");
-        }
-        return "";
-    }, [urlParams, hasCookie]);
-
-    const [cookie, _] = useCookies([cookieParam]);
-
-    if (cookie[cookieParam] && cookie[cookieParam] !== "") {
-        cookieAuth = cookie[cookieParam];
-        cookiesAvailable = true;
-    }
-    return { cookieAuth, cookiesAvailable };
-}
-
-// useUrlAvailable:
-
-export function useUrlAvailable(urlParams: any) {
-    const hasOneForAll = useMemo(() => {
-        return urlParams.has("url");
-    }, [urlParams]);
-
-    const oneForAllParam = useMemo(() => {
-        if (hasOneForAll) {
-            return urlParams.get("url");
-        }
-        return "";
-    }, [urlParams, hasOneForAll]);
-
-    return { url: oneForAllParam, urlAvailable: hasOneForAll };
-}
-
-// updateDataSources:
-
-export function updateDataSourcesWithUrl(
-    dispatch: any,
-    url: any,
-    cookies: any,
-    haveUrl: any,
-    haveCookies: any,
-    dataSources: any
-) {
-    let apiUrl = "";
-    let basicAuth = false;
-    let urlApi = false;
-    let cookieAuth: any = {};
-
-    if (haveUrl) {
-        urlApi = true;
-    }
-
-    if (haveCookies) {
-        let [auth, dsData] = cookies.split("@");
-        let cookieDsData = "";
-        if (dsData && dsData !== "") {
-            try {
-                cookieDsData = atob(dsData);
-                cookieDsData = JSON.parse(cookieDsData);
-                if (typeof cookieDsData === "object" && cookieDsData["url"]) {
-                    apiUrl = cookieDsData["url"];
-                    haveUrl = true;
-                    urlApi = true;
-                }
-            } catch (e) {
-                console.log(e);
-            }
-        }
-
-        let [user, pass] = auth.split(":");
-
-        if (user !== "" && pass !== "") {
-            cookieAuth = { user, password: pass };
-            basicAuth = true;
-        }
-    }
-
-    if (!haveUrl && basicAuth) {
-        apiUrl = window.location.protocol + "//" + window.location.host;
-        urlApi = true;
-    }
-
-    if (apiUrl === "") {
-        urlApi = true;
-        apiUrl = url;
-    }
-
-    const dsCP = [...dataSources];
-    const prevDs = JSON.parse(JSON.stringify(dsCP));
-
-    const newDs = prevDs?.map((m: any) => ({
-        ...m,
-        url: urlApi ? apiUrl : m.url,
-        auth: {
-            ...m.auth,
-            basicAuth: { ...m.auth.basicAuth, value: basicAuth },
-            fields: {
-                ...m.auth.fields,
-                basicAuth: basicAuth
-                    ? [...m.auth.fields.basicAuth]?.map((ba) => {
-                          if (ba.name === "user") {
-                              return { ...ba, value: cookieAuth.user };
-                          }
-                          if (ba.name === "password") {
-                              return { ...ba, value: cookieAuth.password };
-                          }
-                          return ba;
-                      })
-                    : [...m.auth.fields.basicAuth],
-            },
-        },
-    }));
-
-    if (cookies && cookieAuth) {
-        dispatch(setShowDataSourceSetting(false));
-    }
-
-    localStorage.setItem("dataSources", JSON.stringify(newDs));
-    dispatch(setDataSources(newDs));
-}
+import { useMemo, useEffect} from "react";
+import { useNavigate, useLocation } from "react-router-dom";
+import { MobileView } from "./Main/MobileView";
+import { DesktopView } from "./Main/DesktopView";
+import { useCookiesAvailable, useUrlAvailable } from "./Main/hooks";
+import { updateDataSourcesWithUrl, updateDataSourcesFromLocalUrl } from "./Main/helpers";
 
 export default function Main() {
-    UpdateStateFromQueryParams();
 
+
+    const navigate = useNavigate();
     const dataSources = useSelector((store: any) => store.dataSources);
     // get hash from current location
     const { hash } = useLocation();
@@ -296,7 +22,8 @@ export default function Main() {
         return new URLSearchParams(hash.replace("#", ""));
     }, [hash]);
     //
-
+    UpdateStateFromQueryParams();
+    
     const { cookiesAvailable, cookieAuth } = useCookiesAvailable(paramsMemo);
     const { urlAvailable, url } = useUrlAvailable(paramsMemo);
     useEffect(() => {
@@ -307,7 +34,7 @@ export default function Main() {
             cookieAuth !== "" &&
             urlAvailable &&
             urlAvailable !== "";
-
+        // else, take url from location
         if (onlyCookie || onlyUrl || urlAndCookie) {
             // update datasources with url and basic auth
             updateDataSourcesWithUrl(
@@ -318,9 +45,10 @@ export default function Main() {
                 cookiesAvailable,
                 dataSources
             );
+        } else {
+            updateDataSourcesFromLocalUrl(dataSources, dispatch, navigate);
         }
     }, []);
-
 
     useEffect(() => {
         const urlSetting = {
@@ -341,7 +69,9 @@ export default function Main() {
     const isEmbed = useSelector((store: any) => store.isEmbed);
     const theme = useSelector((store: any) => store.theme);
     const autoTheme = useSelector((store: any) => store.autoTheme);
-    const settingsDialogOpen = useSelector((store: any) => store.settingsDialogOpen);
+    const settingsDialogOpen = useSelector(
+        (store: any) => store.settingsDialogOpen
+    );
     const themeMemo = useMemo(() => (themes as any)[theme], [theme]);
 
     useEffect(() => {
